+++ src/Settings.ts
@@ -11,21 +11,24 @@ export class Settings {
     return Settings.configuration.update(key, val, isGlobal);
   }
   static get position() {
-    return (Settings.getSettings('position') as 'Right' | 'Left') || 'Left';
+    return Settings.getSettings('position') as 'Right' | 'Left';
   }
   static get priority() {
     return (Settings.getSettings('priority') as number) || -1;
   }
   static get fileSizeFormat() {
-    return (Settings.getSettings('fileSizeFormat') as string) || '$(file-text) ${fileSize}';
+    return Settings.getSettings('fileSizeFormat') as string;
   }
   static get selectionCountFormat() {
-    return (Settings.getSettings('selectionCountFormat') as string) || '  ${linesCount} : ${wordsCount}';
+    return Settings.getSettings('selectionCountFormat') as string;
   }
   static get dataCountFormat() {
-    return (Settings.getSettings('dataCountFormat') as string) || ' : ${dataCountWithBrackets}';
+    return Settings.getSettings('dataCountFormat') as string;
   }
   static get visibility() {
-    return (Settings.getSettings('visibility') as { fileSize: boolean; selection: boolean; data: boolean }) || {};
+    return Settings.getSettings('visibility') as { fileSize: boolean; selection: boolean; data: boolean };
+  }
+  static set visibility(object: { fileSize: boolean; selection: boolean; data: boolean }) {
+    Settings.setSettings('visibility', object);
   }
 }