+++ src/enum.ts
@@ -1,1 +1,2 @@
 export const SHOW_DETAILS = 'data-size-count.showDetails';
+export const SHOW_HIDE_STATUSBAR = 'data-size-count.showHideStatusBar';